# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API - WA Laboratory",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/exam/actives":{
      "get": {
        "tags":["Exam"],
        "summary": "Remove Exams actives",
        "description": "",
        "produces": [
          "application/json"
        ], 
       "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/ExamArray"
            },
            "description": "Exam found."
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/exam/remove":{
      "post": {
        "tags":["Exam"],
        "summary": "Get all Exams actives",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Exam information.",
            "required": true,
            "schema": {
              "type":"object",
              "properties": {
                "idExam": {
                  "type": "string",
                  "example": "asdasd-23423-sd434sd-ffsdfsd"
                }
            },
            "description": "Exam found."
            }
          }
        ],
       "responses": {
          "200": {
            "schema": {
              "type":"object",
              "properties": {
                "response": {
                  "type": "string",
                  "example": "Exam removed successfully!"
                }
            },
            "description": "Exam found."
          }
        }, "500": {
          "description": "Error on to remove exam!"
        }
      }
    }
    },
    "/api/exam/getLaboratoriesByName?name={name}":{
      "get": {
        "tags":["Exam"],
        "summary": "Get all Laboratories linked withe exam by name exam",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Exam information.",
            "required": true,
            "schema": {
              "type":"object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Exam Name"
                }
            },
            "description": "Exam found."
            }
          }
        ],
      "responses": {
        "200": {
          "schema": {
            "$ref": "#/definitions/LaboratoryArray"
          },
          "description": "Laboratories found."
        }, "500": {
          "description": "Error on to get laboratories that have linked with this exam!"
        }
      }
    }
    },
    "/api/exam/saveBatch":{
      "post": {
        "tags":["Exam"],
        "summary": "Save exams in batch",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Exam information.",
            "required": true,
            "schema": {
                "$ref": "#/definitions/ExamArrayBatch"
              }
          }
        ],
      "responses": {
        "200": {
          "schema": {
            "type":"object",
            "properties": {
              "response": {
                "type": "string",
                "example": "Exams stored successfully!"
              }
          },
          "description": "Exam found."
        }
        }, "500": {
          "description": "Error on stored exams!"
        }
      }
    }
    },
    "/api/exam/deleteBatch":{
      "post": {
        "tags":["Exam"],
        "summary": "Delete exams in batch",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Exam information.",
            "required": true,
            "type": "array",
            "items":{
              "type":"object",
              "properties":
              {
                "ids":
              {
                "type": "array",
                "items":{
                  "type":"string",
                  "example": "asdasd-23423-sd434sd-ffsdfsd"
                }
              }
            }
          }
          }
        ],
      "responses": {
        "200": {
          "schema": {
            "type":"object",
            "properties": {
              "response": {
                "type": "string",
                "example": "Exams deleted successfully!"
              }
          },
          "description": "Exam found."
        }
        }, "500": {
          "description": "Error on delete exams!"
        }
      }
    }
    },
    "/api/exam/updateBatch":{
      "post": {
        "tags":["Exam"],
        "summary": "Update exams in batch",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Exam information.",
            "required": true,
            "schema": {
                "$ref": "#/definitions/ExamArrayBatch"
              }
          }
        ],
      "responses": {
        "200": {
          "schema": {
            "type":"object",
            "properties": {
              "response": {
                "type": "string",
                "example": "Exams updated successfully!"
              }
          },
          "description": "Exam found."
        }
        }, "500": {
          "description": "Error updating exams!"
        }
      }
    }
    },
...
  },"definitions": 
...
    "ExamArray": {
      "type": "array",
      "items":{
        "type":"object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "asdasd-23423-sd434sd-ffsdfsd"
          },
          "name": {
            "type": "string",
            "example": "Exam Name"
          },
          "type": {
            "type": "enum",
            "example": "Image or Clinic or Analysis"
          },
          "status": {
            "type": "enum",
            "example": "Active or Inactive"
          }
        },"required": [
          "name",
          "type",
          "status"
        ],
        "xml": {
          "name": "ExamArray"
        }
      }
    },
    "ExamArrayBatch": {
      "type": "array",
      "items":{
        "type":"object",
        "properties":
        {
          "exams":
        {
          "type": "array",
          "items":{
            "type":"object",
              "properties":{
                "_id": {
                  "type": "string",
                  "example": "asdasd-23423-sd434sd-ffsdfsd"
                },
                "name": {
                  "type": "string",
                  "example": "Exam Name"
                },
                "type": {
                  "type": "enum",
                  "example": "Image or Clinic or Analysis"
                },
                "status": {
                  "type": "enum",
                  "example": "Active or Inactive"
                }
              }
            }
        }
        },"required": [
          "name",
          "type",
          "status"
        ],
        "xml": {
          "name": "ExamArray"
        }
      }
    },
...
    "LaboratoryArray": {
      "type": "array",
      "items":{
        "type":"object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "asdasd-23423-sd434sd-ffsdfsd"
          },
          "name": {
            "type": "string",
            "example": "Laboratory Name"
          },
          "address": {
            "type": "string",
            "example": "Address laboratory"
          },
          "status": {
            "type": "enum",
            "example": "Active or Inactive"
          }
        },"required": [
          "name",
          "address",
          "status"
        ],
        "xml": {
          "name": "LaboratoryArray"
        }
      }
    },
...
  }
}